Stable Release for Beta Testing
I had to change the version numbers back to what Signal had them set to. I want to test later, which ones need to be set to what. It'd be nice if I could use v2.o for HCP, but there was something in the server requests, that required a different version, to register. It may be all of the version, strings, but I don't know.